Refactor: SseServiceImpl 로그 변경 및 테스트 코드의 중복 given 유틸화
