+++ src/test/java/com/stylemycloset/notification/event/listener/FeedLikedNotificationEventListenerIntegrationTest.java
@@ -1,16 +1,16 @@
 package com.stylemycloset.notification.event.listener;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 import static org.testcontainers.shaded.org.awaitility.Awaitility.await;
 
-import com.stylemycloset.notification.TestUserFactory;
 import com.stylemycloset.notification.entity.Notification;
 import com.stylemycloset.notification.entity.NotificationLevel;
 import com.stylemycloset.notification.event.domain.FeedLikedEvent;
 import com.stylemycloset.notification.repository.NotificationRepository;
+import com.stylemycloset.notification.util.NotificationStubHelper;
+import com.stylemycloset.notification.util.TestUserFactory;
 import com.stylemycloset.ootd.entity.Feed;
 import com.stylemycloset.ootd.repo.FeedRepository;
 import com.stylemycloset.sse.repository.SseRepository;
@@ -18,7 +18,6 @@ import com.stylemycloset.sse.service.impl.SseServiceImpl;
 import com.stylemycloset.testutil.IntegrationTestSupport;
 import com.stylemycloset.user.entity.User;
 import com.stylemycloset.user.repository.UserRepository;
-import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -66,13 +65,7 @@ public class FeedLikedNotificationEventListenerIntegrationTest extends Integrati
 
     given(feedRepository.findWithUserById(feed.getId())).willReturn(Optional.of(feed));
     given(userRepository.findById(likeUser.getId())).willReturn(Optional.of(likeUser));
-    given(notificationRepository.save(any(Notification.class)))
-        .willAnswer(invocation -> {
-          Notification n = invocation.getArgument(0);
-          ReflectionTestUtils.setField(n, "id", 1L);
-          ReflectionTestUtils.setField(n, "createdAt", Instant.now());
-          return n;
-        });
+    NotificationStubHelper.stubSave(notificationRepository);
     given(sseRepository.findByUserId(user.getId())).willReturn(new CopyOnWriteArrayList<>(List.of(emitter)));
 
     FeedLikedEvent event = new FeedLikedEvent(6L, 66L);
